feat:add stake sol functionality

+++ src/instructions.rs
@@ -3,7 +3,7 @@ use borsh::{BorshDeserialize, BorshSerialize};
 #[derive(BorshDeserialize , BorshSerialize)]
 pub enum StakingInstruction {
 InitializePool{reward_rate : u64},
-Stake {amount : u32},
-Unstake{amount :u32},
+Stake {amount : u64},
+Unstake{amount :u64},
 ClaimRewards,
 }
